refactor persona info
