--- src/pages/persona/[persona].ts
@@ -1,2 +1,0 @@
-import PersonaInfo from "../../persona";
-export default PersonaInfo;
